Move invoker to its own package.
