+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/invoker/ICommandFactory.java
@@ -1,7 +1,9 @@
-package org.metaborg.spoofax.shell.commands;
+package org.metaborg.spoofax.shell.invoker;
 
 import org.metaborg.core.language.ILanguageImpl;
 import org.metaborg.core.project.IProject;
+import org.metaborg.spoofax.shell.commands.AnalyzeCommand;
+import org.metaborg.spoofax.shell.commands.ParseCommand;
 
 /**
  * Factory for creating Spoofax commands.
